+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   for SA-AADC VM.
 - `dsclibrary\MEMBER_AADC.DSC.ps1`: Created DSC config for deploying an
   Azure AD Connect server.
+- `dsclibrary\MEMBER_WAC.DSC.ps1`: Created DSC config for deploying a
+  Windows Administration Center server.
 
 ## [1.1.0] - 2020-08-30